requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
SQLAlchemy==1.4.26
Werkzeug==2.0.2
pymysql==1.0.2
python-dotenv==0.20.0